+++ app/Http/Controllers/CotacoesController.php
@@ -5,6 +5,7 @@ namespace Softage\Http\Controllers;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
 
+use Illuminate\Support\Facades\Auth;
 use Softage\Repositories\CotacoesRepository;
 use Softage\Services\CotacoesService;
 use Softage\Http\Requests;
@@ -92,14 +93,26 @@ class CotacoesController extends Controller
         #Editando a grid
         return Datatables::of($rows)->addColumn('action', function ($row) {
             # Html de retorno
-            $html = '<a href="edit/'.$row->id.'" class="btn btn-xs btn-primary"><i class="glyphicon glyphicon-edit"></i> Editar</a> ';
+            $html = "";
+
+            # Recuperando o usuário;
+            $user = Auth::user();
+
+            # Checando permissão
+            if($user->can('cotacao.update')) {
+                # Html de retorno
+                $html = '<a href="edit/' . $row->id . '" class="btn btn-xs btn-primary"><i class="glyphicon glyphicon-edit"></i> Editar</a> ';
+            }
 
             # Recuperando a cotação
             $cotacao = $this->repository->find($row->id);
             
             # Validando a possibilidade de remoção
             if(!count($cotacao->apostas) > 0) {
-                $html .= '<a href="destroy/'.$row->id.'" class="btn btn-xs btn-danger delete"><i class="glyphicon glyphicon-delete"></i> Remover</a>';
+                # Checando permissão
+                if($user->can('cotacao.destroy')) {
+                    $html .= '<a href="destroy/' . $row->id . '" class="btn btn-xs btn-danger delete"><i class="glyphicon glyphicon-delete"></i> Remover</a>';
+                }
             }
 
             # retorno
