+++ app/Http/Controllers/ModalidadesController.php
@@ -4,6 +4,7 @@ namespace Softage\Http\Controllers;
 
 use Illuminate\Http\Request;
 
+use Illuminate\Support\Facades\Auth;
 use Softage\Repositories\ModalidadesRepository;
 use Softage\Services\ModalidadesService;
 use Softage\Http\Requests;
@@ -80,14 +81,26 @@ class ModalidadesController extends Controller
         #Editando a grid
         return Datatables::of($rows)->addColumn('action', function ($row) {
             # Html de retorno
-            $html = '<a href="edit/'.$row->id.'" class="btn btn-xs btn-primary"><i class="glyphicon glyphicon-edit"></i> Editar</a> ';
+            $html = "";
+
+            # Recuperando o usuário;
+            $user = Auth::user();
+
+            # Checando permissão
+            if($user->can('modalidade.update')) {
+                # Html de retorno
+                $html = '<a href="edit/' . $row->id . '" class="btn btn-xs btn-primary"><i class="glyphicon glyphicon-edit"></i> Editar</a> ';
+            }
 
             # Recuperando a modalidade
             $mdalidade = $this->repository->find($row->id);
 
             # Validando a possibilidade de remoção
             if(!count($mdalidade->cotacoes) > 0) {
-                $html .= '<a href="destroy/'.$row->id.'" class="btn btn-xs btn-danger delete"><i class="glyphicon glyphicon-delete"></i> Remover</a>';
+                # Checando permissão
+                if($user->can('modalidade.destroy')) {
+                    $html .= '<a href="destroy/' . $row->id . '" class="btn btn-xs btn-danger delete"><i class="glyphicon glyphicon-delete"></i> Remover</a>';
+                }
             }
 
             # retorno
