+++ app/Http/Controllers/TimesController.php
@@ -59,12 +59,24 @@ class TimesController extends Controller
 
         #Editando a grid
         return Datatables::of($rows)->addColumn('action', function ($row) {
+            # Html de retorno
             $html = "";
-            $html .= '<a href="edit/'.$row->id.'" class="btn btn-xs btn-primary"><i class="glyphicon glyphicon-edit"></i> Editar</a> ';
+
+            # Recuperando o usuário;
+            $user = Auth::user();
+
+            # Checando permissão
+            if($user->can('times.update')) {
+                $html .= '<a href="edit/' . $row->id . '" class="btn btn-xs btn-primary"><i class="glyphicon glyphicon-edit"></i> Editar</a> ';
+            }
+
             # Verificando se existe vinculo com o currículo
             $time = $this->service->find($row->id);
             if(count($time->partidas) == 0) {
-                $html .= '<a href="delete/'.$row->id.'" class="btn btn-xs btn-danger delete"><i class="glyphicon glyphicon-edit"></i> Deletar</a>';
+                # Checando permissão
+                if($user->can('times.destroy')) {
+                    $html .= '<a href="delete/' . $row->id . '" class="btn btn-xs btn-danger delete"><i class="glyphicon glyphicon-edit"></i> Deletar</a>';
+                }
             }
             return $html;
         })->make(true);
